requirements_enhanced.txt
# Enhanced OMR requirements
tensorflow>=2.8.0
scikit-learn>=1.2.0
numpy>=1.21.0
opencv-python-headless>=4.5.3.56
matplotlib>=3.5.0
pillow>=8.3.0
scipy>=1.7.0
types-Pillow
types-tensorflow
typing-extensions
onnxruntime-gpu
pandas>=1.3.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0